### 2022-02-08 #503 @joeroe
rpaleoclim
Submitting Author Name: Joe Roe
Submitting Author Github Handle: @joeroe
Repository:  https://github.com/joeroe/rpaleoclim
Version submitted: 0.9.0.9000
Submission type: Standard

Version accepted: TBD
---

```
Package: rpaleoclim
Title: R Interface for the PaleoClim Paleoclimate Database
Version: 0.9.0.9000
    person(given = "Joe",
           family = "Roe",
           email = "[joe@joeroe.io](mailto:joe@joeroe.io)")
    PaleoClim <http://paleoclim.org> is a database of free, high resolution
  paleoclimate surfaces covering the whole globe. It includes the standard
  bioclimatic variables commonly used in ecological modelling, downscaled from
  the HadCM3 general circulation model for key time periods, at up to 2.5 minute
  spatial resolution.
    This package provides a simple interface for downloading PaleoClim data in R,
  with support for spatial queries and caching.
    fs,
    rgdal,
    raster,
    curl
    covr,
    mockery,
```

rpaleoclim retrieves data from [PaleoClim](http://www.paleoclim.org/), a set of high-resolution global paleoclimate surfaces.

As the PaleoClim authors put it, "easily accessible paleoclimate data are essential for environmental, evolutionary, and ecological studies". I would add to that studies of human ecology in the past, e.g. environmental archaeology. The package is simple but makes it easier to reproducibly use this data in R analyses.

There are packages for retrieving *current* global climate data that are widely used, for example [WorldClim](https://worldclim.org/) is available via the `raster` package, but I'm not aware of any packages for accessing paleoclimate data, or PaleoClim specifically.

Yes. No human subjects or personal information is involved. I have endeavoured to 'politely' make requests to [SDMtoolbox](http://www.sdmtoolbox.org/)'s server (see https://github.com/joeroe/rpaleoclim/issues/6), which currently hosts the PaleoClim data, and the caching function reduces server load.

pkgcheck results: https://github.com/joeroe/rpaleoclim/actions/runs/1813065454

Regarding documentation: functions are documented, but I have removed examples because this causes persistent problems with server timeouts when they are run on CI. This is frustrating because the file downloaded is not very big (16MB) and I have not been able to reproduce these problems with the SDMtoolbox server locally. It is not really feasible to write examples that do not download data or rely on files that are too large to be on CRAN. I realise this is not ideal; any advice on how to get around this problem would be much appreciated!

There is no vignette because the package has a very simple API (two functions) and I think this is adequately explained in the README.
