### 2022-02-09 b659fa7
Switch from terra to raster #10
Providing a temporary `as = "raster" argument for backwards compatibility. To be removed in future versions.
