+++ man/load_paleoclim.Rd
@@ -4,14 +4,21 @@
 \alias{load_paleoclim}
 \title{Load data from PaleoClim}
 \usage{
-load_paleoclim(file)
+load_paleoclim(file, as = c("terra", "raster"))
 }
 \arguments{
 \item{file}{Character. Path to a *.zip file downloaded from PaleoClim.}
+
+\item{as}{Character. \code{as = "raster"} returns a \code{RasterStack} object
+(see \code{\link[raster:stack]{raster::stack()}}) instead of the default raster from
+the \code{terra} package. It is provided for backwards
+compatibility and will be removed in future versions.
+Requires the \code{raster} package.}
 }
 \value{
-\code{RasterStack} object (see \code{\link[raster:stack]{raster::stack()}}).
+\code{SpatRaster} object (see \code{\link[terra:rast]{terra::rast()}}) with each bioclimatic variable
+as a separate named layer.
 }
 \description{
-Loads a PaleoClim data file (\code{.zip} format) into R as a \code{RasterStack} object.
+Loads a PaleoClim data file (\code{.zip} format) into R as a \code{SpatRaster}.
 }
