+++ README.md
@@ -39,11 +39,12 @@ remotes::install_github("joeroe/rpaleoclim")
 The package includes two functions, `paleoclim()` and
 `load_paleoclim()`. `paleoclim()` downloads data from PaleoClim with the
 desired period and resolution and reads it into R as a
-[raster](https://rspatial.org/raster/) object.
+[SpatRaster](https://rspatial.org/spatial-terra/8-rastermanip.html)
+object.
 
 ``` r
 library("rpaleoclim")
-library("raster")
+library("terra") # For plotting
 
 paleoclim("lh", "10m") |>
   plot()
