+++ README.Rmd
@@ -38,11 +38,11 @@ remotes::install_github("joeroe/rpaleoclim")
 ## Usage
 
 The package includes two functions, `paleoclim()` and `load_paleoclim()`.
-`paleoclim()` downloads data from PaleoClim with the desired period and resolution and reads it into R as a [raster](https://rspatial.org/raster/) object.
+`paleoclim()` downloads data from PaleoClim with the desired period and resolution and reads it into R as a [SpatRaster](https://rspatial.org/spatial-terra/8-rastermanip.html) object.
 
 ```{r rpaleoclim-demo, message=FALSE}
 library("rpaleoclim")
-library("raster")
+library("terra") # For plotting
 
 paleoclim("lh", "10m") |>
   plot()

[…]
