### 2022-02-09 ff9d998
Require rgdal for backwards compatibility too
