+++ R/paleoclim.R
@@ -13,7 +13,7 @@
 #'                    (see [raster::stack()]) instead of the default raster from
 #'                    the `terra` package. It is provided for backwards
 #'                    compatibility and will be removed in future versions.
-#'                    Requires the `raster` package.
+#'                    Requires the `raster` and `rgdal` packages.
 #' @param skip_cache  Logical. If `TRUE`, cached data will be ignored.
 #' @param cache_path  Logical. Path to directory where downloaded files should
 #'   be saved. Defaults to R's temporary directory.
@@ -126,7 +126,7 @@ construct_paleoclim_url <- function(period, resolution) {
 #'                    (see [raster::stack()]) instead of the default raster from
 #'                    the `terra` package. It is provided for backwards
 #'                    compatibility and will be removed in future versions.
-#'                    Requires the `raster` package.
+#'                    Requires the `raster` and `rgdal` packages.
 #'
 #' @return
 #' `SpatRaster` object (see [terra::rast()]) with each bioclimatic variable
@@ -145,9 +145,10 @@ load_paleoclim <- function(file, as = c("terra", "raster")) {
   raster <- terra::rast(tifs)
 
   if (as == "raster") {
-    if (!requireNamespace("raster", quietly = TRUE)) {
-      rlang::error(
-        '`as = "raster"` requires package `raster`'
+    if (!requireNamespace("raster", quietly = TRUE) ||
+        !requireNamespace("rgdal", quietly = TRUE)) {
+      rlang::abort(
+        '`as = "raster"` requires packages `raster` and `rgdal`'
       )
     }
 
