+++ DESCRIPTION
@@ -31,5 +31,6 @@ Suggests:
     covr,
     mockery,
     raster,
+    rgdal,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
