### 2022-02-11 ddd1cb7
Try paleoclim() example with \donttest again
