+++ man/paleoclim.Rd
@@ -47,3 +47,8 @@ temporary directory if available. Use \code{skip_cache = TRUE} to override this.
 \code{cache_path} can also be set to another directory. This can be useful if you
 want to reuse downloaded data between sessions.
 }
+\examples{
+\donttest{
+paleoclim("lh", "10m")
+}
+}
