### 2022-02-11 6102de6
Don't run \donttest examples in R CMD check workflow
