+++ .github/workflows/R-CMD-check.yaml
@@ -27,6 +27,7 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
+      R_CHECK_DONTTEST_EXAMPLES: false
 
     steps:
       - uses: actions/checkout@v2
