### 2022-02-11 7be3575
Add example for load_paleoclim()
