+++ man/load_paleoclim.Rd
@@ -15,3 +15,8 @@ load_paleoclim(file)
 \description{
 Loads a PaleoClim data file (\code{.zip} format) into R as a \code{RasterStack} object.
 }
+\examples{
+file <- system.file("testdata", "LH_v1_10m_cropped.zip",
+                    package = "rpaleoclim")
+load_paleoclim(file)
+}
