### 2022-02-11 b3e4656
Revert addition of R_CHECK_DONTTEST_EXAMPLES
Doesn't seem to have any effect.
