### 2022-02-11 a8aa4e1
Use timeout option in curl_download()
Default is 60 so hopefully will help with #11. And providing extra options for people doesn't hurt
