### 2022-02-11 d843935
Fix mock_download()
Also add it to crop test
