### 2022-02-11 f9e400c
Increase timeout limit further #11
Since it should be enough to download the whole file. The largest datasets from PaleoClim are 3.6 GB, so go high.
