### 2022-02-16 fcc3f5d
Update NEWS.md
