### 2022-02-16 73df346
Check we can resolve sdmtoolbox.org before downloading
So we can at least get a slightly more informative error message on server problems
