### 2022-02-16 6ab7572
Mark paleoclim() example \dontrun #11
