### 2022-02-16 5b2c31f
Don't use pipes
