### 2022-02-16 33e2181
Require raster >= 3.0-2
Otherwise there's no method for converting SpatRaster to raster
