+++ README.Rmd
@@ -13,7 +13,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# rpaleoclim: an R interface for *PaleoClim* paleoclimate data
+# rpaleoclim: download paleoclimate data from PaleoClim in R
 
 <!-- badges: start -->
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
@@ -23,10 +23,10 @@ knitr::opts_chunk$set(
 [![DOI](https://zenodo.org/badge/199434717.svg)](https://zenodo.org/badge/latestdoi/199434717)
 <!-- badges: end -->
 
-[PaleoClim](http://paleoclim.org) is a set of free, high resolution paleoclimate surfaces covering the whole globe.
-It includes the standard bioclimatic variables commonly used in ecological modelling, downscaled from the HadCM3 general circulation model for key time periods, at up to 2.5 minute spatial resolution.
+[PaleoClim](http://paleoclim.org) is a set of free, high resolution paleoclimate surfaces covering the whole globe. 
+It includes data on surface temperature, precipitation and the standard bioclimatic variables commonly used in  ecological modelling, derived from the 'HadCM3' general circulation model and  downscaled to a spatial resolution of up to 2.5 minutes. 
 
-This package provides a simple interface for downloading PaleoClim data in R, with support for spatial queries and caching.
+This package provides a simple interface for downloading 'PaleoClim' data in R, with support for caching and filtering retrieved data by period, resolution, and geographic extent.
 
 ## Installation
 
