+++ man/rpaleoclim-package.Rd
@@ -4,9 +4,9 @@
 \name{rpaleoclim-package}
 \alias{rpaleoclim}
 \alias{rpaleoclim-package}
-\title{rpaleoclim: R Interface for the PaleoClim Paleoclimate Database}
+\title{rpaleoclim: Download Paleoclimate Data from 'PaleoClim'}
 \description{
-PaleoClim <http://paleoclim.org> is a database of free, high resolution paleoclimate surfaces covering the whole globe. It includes the standard bioclimatic variables commonly used in ecological modelling, downscaled from the HadCM3 general circulation model for key time periods, at up to 2.5 minute spatial resolution. This package provides a simple interface for downloading PaleoClim data in R, with support for spatial queries and caching.
+'PaleoClim' <http://paleoclim.org> (Brown et al. 2019, <doi:10.1038/sdata.2018.254>) is a set of free, high resolution paleoclimate surfaces covering the whole globe. It includes data on surface temperature, precipitation and the standard bioclimatic variables commonly used in ecological modelling, derived from the 'HadCM3' general circulation model and downscaled to a spatial resolution of up to 2.5 minutes. Simulations are available for key time periods from the Late Holocene to mid-Pliocene. Data on current and Last Glacial Maximum climate is derived from 'CHELSA' (Karger et al. 2017, <doi:10.1038/sdata.2017.122>) and reprocessed by 'PaleoClim' to match their format; it is available at up to 30 seconds resolution. This package provides a simple interface for downloading 'PaleoClim' data in R, with support for caching and filtering retrieved data by period, resolution, and geographic extent.
 }
 \seealso{
 Useful links:
