+++ vignettes/rpaleoclim.Rmd
@@ -1,0 +1,139 @@
+---
+title: "Introduction to rpaleoclim"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{rpaleoclim}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r rmarkdown-options, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+[PaleoClim](http://paleoclim.org) is a set of free, high resolution paleoclimate surfaces covering the whole globe. 
+It includes data on surface temperature, precipitation and the standard bioclimatic variables commonly used in  ecological modelling, derived from the HadCM3 general circulation model and  downscaled to a spatial resolution of up to 2.5 minutes.
+Simulations are available for key time periods from the Late Holocene to mid-Pliocene. 
+Data on current and Last Glacial Maximum climate is derived from [CHELSA](https://chelsa-climate.org/) and reprocessed by PaleoClim to match their format; it is available at up to 30 seconds resolution.
+
+This package provides a simple interface for downloading PaleoClim data in R, with support for caching and filtering retrieved data by period, resolution, and geographic extent.
+
+# Installation
+
+You can install the development version of rpaleoclim from GitHub using the [`remotes`](https://github.com/r-lib/remotes) package:
+
+```{r install-rpaleoclim, eval=FALSE}
+remotes::install_github("joeroe/rpaleoclim")
+```
+
+It depends on `terra`, which in turn requires a recent version of the the system libraries [GDAL](https://gdal.org/) and [PROJ](https://proj.org/).
+These are included in the binary releases of `terra` for Windows and MacOS, but if you build from source (e.g. on Linux) you might need to install them first;
+see the [terra README](https://github.com/rspatial/terra#from-source-code) for instructions for different systems.
+
+# Data available
+
+rpaleoclim provides an R interface to download all the data listed at <http://www.paleoclim.org/> and currently mirrored at <http://www.sdmtoolbox.org>.
+The tables below were last updated 2022-02-16; please refer to the PaleoClim website for the authoritative version.
+
+If you notice a change or update to the PaleoClim data structure that isn't supported by the current version of the package, please report it at: <https://github.com/joeroe/rpaleoclim/issues>.
+
+## Periods
+
+PaleoClim includes climate reconstructions from simulations of the following time intervals, supplemented with two additional datasets from CHELSA:
+
+```{r paleoclim-periods, echo=FALSE}
+knitr::kable(read.csv("paleoclim_periods.csv"))
+```
+
+## Bioclimatic variables
+
+## Other options
+
+The options for `resolution` are:
+
+* `"10m"`: 10 minutes, c. 16 km at 30° N/S
+* `"5m"`: 5 minutes, c. 8 km at 30° N/S
+* `"2_5m"`: 2.5 minutes, c. 4 km at 30° N/S
+* `"30s"`: 30 seconds, c. 1 km at 30° N/S (only available for `"cur"` and `"lgm"`)
+
+## Citations
+
+Please follow the [instructions from the authors](http://paleoclim.org/how-to-cite/) when citing PaleoClim data. 
+At time of writing, this includes a citation to the paper the describing the PaleoClim database:
+
+* Brown, J.L., Hill, D.J., Dolan, A.M., Carnaval, A.C., Haywood, A.M., 2018. [PaleoClim, high spatial resolution paleoclimate surfaces for global land areas](https://www.nature.com/articles/sdata2018254). *Scientific Data* 5, 180254. doi:10.1038/sdata.2018.254
+
[…]
