+++ README.Rmd
@@ -24,9 +24,9 @@ knitr::opts_chunk$set(
 <!-- badges: end -->
 
 [PaleoClim](http://paleoclim.org) is a set of free, high resolution paleoclimate surfaces covering the whole globe. 
-It includes data on surface temperature, precipitation and the standard bioclimatic variables commonly used in  ecological modelling, derived from the 'HadCM3' general circulation model and  downscaled to a spatial resolution of up to 2.5 minutes. 
+It includes data on surface temperature, precipitation and the standard bioclimatic variables commonly used in  ecological modelling, derived from the HadCM3 general circulation model and  downscaled to a spatial resolution of up to 2.5 minutes. 
 
-This package provides a simple interface for downloading 'PaleoClim' data in R, with support for caching and filtering retrieved data by period, resolution, and geographic extent.
+This package provides a simple interface for downloading PaleoClim data in R, with support for caching and filtering retrieved data by period, resolution, and geographic extent.
 
 ## Installation
 
