### 2022-03-11 #503 @jooolia
Dear @joeroe,
Will you pursue fixing the two point mentioned above? Thanks, Julia
