### 2022-03-12 #503 @joeroe
Sorry for the delay and thanks for the feedback @jooolia. I've added a vignette and examples for all functions, but I'm still struggling with the server errors that sometimes cause the package to fail R CMD check. I have a couple of things I still want to try to fix it; hopefully early next week.
