### 2022-05-11 e236f4b
Always quieten paleoclim() in non-interactive mode
Mocking tests accordingly
