### 2022-05-11 #503 @joeroe
Thank you for your patience @jooolia. I can't get around the server issues, so I have contacted the authors of the data to ask about mirroring it somewhere else. Would it be possible to keep this open just a little longer, while I wait for a response?
