+++ R/paleoclim.R
@@ -40,7 +40,7 @@
 #' paleoclim("lh", "10m")
 #' }
 paleoclim <- function(period = c("lh", "mh", "eh", "yds", "ba", "hs1",
-                                 "lig", "mis", "mpwp", "m2", "cur", "lgm"),
+                                 "lig", "mis19", "mpwp", "m2", "cur", "lgm"),
                       resolution = c("10m", "5m", "2_5m", "30s"),
                       region = NULL,
                       as = c("terra", "raster"),
