### 2022-06-29 #503 @joeroe
Still no response unfortunately. I've also recently learned of [pastclim](https://www.biorxiv.org/content/10.1101/2022.05.18.492456v1), which likely makes this package obsolete.

With that in mind I'll withdraw this submission – sorry for taking up your time unnecessarily.
